Separate Flutter from library code
