+++ spectrum/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:scoped_model/scoped_model.dart';
 import 'spectrum_model.dart';
 import 'spectrum_display.dart';
-import 'utility.dart';
+import 'zxspectrum/utility.dart';
 
 void main() => runApp(ProjectCambridge());
