+++ spectrum/lib/zxspectrum/memory.dart
