+++ spectrum/lib/zxspectrum/spectrumcolor.dart
