+++ spectrum/lib/zxspectrum/utility.dart
